Add timeouts to pause a running match
Divide gametime client- and serverside into realTime and gameTime.
realTime corresponds with the serverTime and is used for the usercommands
and client communication. gameTime is stopped during timeout and all
traces and entitystates are calculated using the gameTime.
Add bool to pmove_t to stop all client calculation during timeout.